--- packages/pcgl-validation/package.json
@@ -1,19 +1,0 @@
-{
-	"name": "@pcgl-daco/validation",
-	"private": true,
-	"version": "0.0.0",
-	"description": "PCGL-DACO Validation",
-	"license": "ISC",
-	"exports": "./dist/index.js",
-	"types": "dist/index.d.ts",
-	"type": "module",
-	"scripts": {
-		"build": "tsc"
-	},
-	"dependencies": {
-		"zod": "^3.24.1"
-	},
-	"devDependencies": {
-		"typescript": "^5.5.3"
-	}
-}
