--- packages/pcgl-validation/src/schemas.ts
@@ -1,78 +1,0 @@
-/*
- * Copyright (c) 2025 The Ontario Institute for Cancer Research. All rights reserved
- *
- * This program and the accompanying materials are made available under the terms of
- * the GNU Affero General Public License v3.0. You should have received a copy of the
- * GNU Affero General Public License along with this program.
- *  If not, see <http://www.gnu.org/licenses/>.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
- * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
- * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
- * SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
- * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
- * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
- * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-import { z } from 'zod';
-import { EmptyOrOptionalString, NonEmptyString, OptionalURLString, TrimmedString } from './common/strings.js';
-import { ONLY_ALPHANUMERIC } from './utils/regex.js';
-
-// Applicant Information Form Section
-export type ApplicantInformationSchemaType = z.infer<typeof applicantInformationSchema>;
-// Institutional Representative Form Section
-export type InstitutionalRepSchemaType = z.infer<typeof institutionalRepSchema>;
-// Project Info Form Section
-export type ProjectInformationSchemaType = z.infer<typeof projectInformationSchema>;
-
-export const applicantInformationSchema = z.object({
-	applicantTitle: NonEmptyString,
-	applicantFirstName: NonEmptyString,
-	applicantMiddleName: EmptyOrOptionalString,
-	applicantLastName: NonEmptyString,
-	applicantSuffix: EmptyOrOptionalString,
-	applicantPrimaryAffiliation: NonEmptyString,
-	applicantInstituteAffiliation: NonEmptyString.email(),
-	applicantProfileUrl: NonEmptyString.url(),
-	applicantPositionTitle: NonEmptyString,
-	institutionCountry: NonEmptyString,
-	institutionState: NonEmptyString,
-	institutionCity: NonEmptyString,
-	institutionPostalCode: NonEmptyString.regex(ONLY_ALPHANUMERIC),
-	institutionStreetAddress: NonEmptyString,
-	institutionBuilding: EmptyOrOptionalString,
-});
-
-export const institutionalRepSchema = z.object({
-	institutionalTitle: NonEmptyString,
-	institutionalFirstName: NonEmptyString,
-	institutionalMiddleName: EmptyOrOptionalString,
-	institutionalLastName: NonEmptyString,
-	institutionalSuffix: EmptyOrOptionalString,
-	institutionalPrimaryAffiliation: NonEmptyString,
-	institutionalInstituteAffiliation: NonEmptyString.email(),
-	institutionalProfileUrl: NonEmptyString.url(),
-	institutionalPositionTitle: NonEmptyString,
-	institutionCountry: NonEmptyString,
-	institutionState: NonEmptyString,
-	institutionCity: NonEmptyString,
-	institutionStreetAddress: NonEmptyString,
-	institutionPostalCode: NonEmptyString.regex(ONLY_ALPHANUMERIC),
-	institutionBuilding: EmptyOrOptionalString,
-});
-
-export const projectInformationSchema = z.object({
-	projectTitle: NonEmptyString,
-	projectWebsite: OptionalURLString,
-	projectBackground: TrimmedString.min(100).max(200),
-	projectAims: TrimmedString.min(100).max(200),
-	projectDataUse: TrimmedString.min(100).max(200),
-	projectMethodology: TrimmedString.min(100).max(200),
-	projectLaySummary: TrimmedString.min(100).max(200),
-	relevantPublicationURL1: NonEmptyString.url(),
-	relevantPublicationURL2: NonEmptyString.url(),
-	relevantPublicationURL3: NonEmptyString.url(),
-});
